Refactor : Repository-wrapper for User
